adicionado documentação e melhorias na organização dos arquivos
