+++ Cargo.toml
@@ -40,6 +40,7 @@ iced.workspace = true
 icon = []
 lucide_icons = ["icon"]
 material_icons = ["icon"]
+toasts = []
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
